chore(nuxt): improve playground
